creating food component
