+++ client/main.js
@@ -3,7 +3,8 @@ const { EVENT_TYPES } = require('./enums/events');
 
 const SnakePlayer = require('./components/player');
 const Game = require('./components/game');
-const SnakeController = require('./controller/snake_controller');
+const SnakeController = require('./controller/controller');
+const FoodComponent = require('./components/food');
 
 const Event = require('./event/event');
 
@@ -39,7 +40,6 @@ const playerManual = new SnakePlayer({
     initialPositionX: 0,
     initialPositionY: 4 * LINK_WIDTH,
 });
-const players = [playerManual, playerAutomated];
 
 const controllerAutomated = new SnakeController(playerAutomated);
 const controllerManual = new SnakeController(playerManual);
@@ -49,9 +49,10 @@ const contollers = [controllerManual, controllerAutomated];
 Event.addListener(EVENT_TYPES.UPDATE_FOOD, () => console.log('update food event triggered...'))
 Event.addListener(EVENT_TYPES.GAME_STARTED, () => console.log('game started ...'))
 
+FoodComponent.setParams({ ctx, foodImageSrc: 'images/link.png' })
+
 const game = new Game({
     ctx, 
-    foodImageSrc: 'images/link.png',
     contollers
 });
 game.start();
